refactor: convert memory/tools.py to memory/tools/ package
